pyproject.toml
[project]
name = "reddit-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpraw>=7.8.1",
    "mcp[cli]>=1.9.3",
    "python-dotenv>=1.1.0",
]
